refactor(neovim): turned configuration into vim plugin
