+++ home-manager/config/nvim/init.lua
@@ -1,0 +1,1 @@
+require 'TheAltF4Stream'.init()
